Extract condition of a partner requiring verification.
Not strictly a partner class but needs to go somewhere.
